pyproject.toml
[tool.poetry]
name = "byoe"
version = "0.1.0"
description = "Bring your own environment"
authors = ["Brendan Moloney <moloney.brendan@gmail.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = ">3.9,<4.0"
typer = ">0.9.0"
sh = ">2.0.6"
pyyaml = ">6.0.1"
rich = ">13.5.2"
requests = "^2.31.0"

[tool.poetry.scripts]
byoe = 'byoe.cli:cli'

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
mypy = "^1.15.0"
types-pyyaml = "^6.0.12.20241230"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
